# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Puraani Jeans APIs",
        "description": "Puraani Jeans API documentation",
        "contact": {
            "email": "developer@puranijeans.com"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": ""
    },
    "servers": [
        {
            "url": "http://puranijeans.test/",
            "description": "dev"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "User Register",
                "description": "User Register here",
                "operationId": "Register",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "country",
                                    "email_address",
                                    "first_name",
                                    "last_name",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "country": {
                                        "type": "numeric"
                                    },
                                    "email_address": {
                                        "type": "email"
                                    },
                                    "first_name": {
                                        "type": "text"
                                    },
                                    "last_name": {
                                        "type": "text"
                                    },
                                    "password": {
                                        "type": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "password"
                                    },
                                    "subscription": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "User Login",
                "description": "Login User Here",
                "operationId": "authLogin",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "user_access_token": {
                "type": "apiKey",
                "description": "user access token",
                "name": "Authorization",
                "in": "header"
            },
            "client_id": {
                "type": "apiKey",
                "description": "client id",
                "name": "x-client-id",
                "in": "header"
            },
            "client_secret": {
                "type": "apiKey",
                "description": "client secret",
                "name": "x-client-token",
                "in": "header"
            }
        }
    }
}